BOAC-5042, initial refactor of store:curated-group
